requirements.txt
Django>=3.2,<3.9
wagtail>=2.13,<2.14
djangorestframework>=3.12.2
django-cors-headers>=3.1.7
djangorestframework-simplejwt>=4.6.0
psycopg2-binary>=2.8
django-extensions>=3.1.0
django-allauth>=0.44.0
wagtailmenus>=3.0.2
django-redis>=4.12.1
